+++ test/integration/resources/app.py
@@ -1,0 +1,8 @@
+import flask
+
+app = flask.Flask(__name__)
+
+
+@app.route("/about")
+def route_about():
+    return "About this application"